+++ src/main/java/io/github/jhooc77/objcubedcompat/mixin/sodium/ShaderLoaderMixin.java
@@ -1,14 +1,15 @@
 package io.github.jhooc77.objcubedcompat.mixin.sodium;
 
 import io.github.jhooc77.objcubedcompat.shader.ObjCubedShaderInjector;
-import net.caffeinemc.mods.sodium.client.gl.shader.ShaderLoader;
 import net.minecraft.resources.Identifier;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(value = ShaderLoader.class, remap = false)
+@Pseudo
+@Mixin(targets = "net.caffeinemc.mods.sodium.client.gl.shader.ShaderLoader", remap = false)
 public abstract class ShaderLoaderMixin {
     @Inject(method = "getShaderSource", at = @At("RETURN"), cancellable = true)
     private static void objCubed$injectNativeTerrain(